add initial tests.
