+++ build.gradle
@@ -20,8 +20,9 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter'
 	implementation 'org.web3j:core:4.8.7'
 	implementation group: 'com.squareup.okhttp3', name: 'okhttp', version: '4.9.3'
-	implementation 'com.github.ipld:java-cid:1.3.1'
+	implementation group: 'com.github.ipld', name: 'java-cid', version: '1.3.1'
 	implementation group: 'com.github.ipfs', name: 'java-ipfs-http-client', version: '1.3.3'
+	implementation group: 'com.google.guava', name: 'guava', version: '31.1-jre'
 
 	// test
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
